fix: handle locale setting error
One of the tests tries to set the locale to `de_DE.cp1252` but when this
is not available on the system an error is raised. Since this is called
during test discovery, the entire test suite errors out. The exception
is caught and instead the test module is skipped.